requirements.txt
discord.py==0.16.5
aiohttp==1.0.5
beautifulsoup4==4.5.3
lxml==3.7.2
python-wowapi==1.0
requests==2.11.1